Changed all int and some float things in Reagent code to Decimals

+++ Content.Server/Chemistry/Metabolism/DefaultDrink.cs
@@ -28,7 +28,7 @@ namespace Content.Server.Chemistry.Metabolism
         }
 
         //Remove reagent at set rate, satiate thirst if a ThirstComponent can be found
-        int IMetabolizable.Metabolize(IEntity solutionEntity, string reagentId, float tickTime)
+        decimal IMetabolizable.Metabolize(IEntity solutionEntity, string reagentId, float tickTime)
         {
             int metabolismAmount = (int)Math.Round(MetabolismRate * tickTime);
             if (solutionEntity.TryGetComponent(out ThirstComponent thirst))
